linux/special-pages/dump_kernel_pages: be resilient to SIGBUS
Gentoo Hardened kernel does not seem to support reading the 2nd page of
vvar zone.  strace reports:

    --- SIGBUS {si_signo=SIGBUS, si_code=BUS_ADRERR, si_addr=0x31480c40000} ---
    +++ killed by SIGBUS +++
    Bus error

with:

    31480c3f000-31480c41000 r--p 00000000 00:00 0       [vvar]

As this signal is raised in a call to ctypes.memmove, there is no easy
way to handle this signal in pure Python.  Instead, use write() function
to dump the pages and keep memmove() to kernel pages, as write() cannot
read them.