machines/archlinux: fix Docker build on Arch Linux
tar is complaining when extracting Arch Linux iso install:

    tar: root.x86_64/etc/xdg/systemd: Directory renamed before its status could be extracted
    tar: root.x86_64/etc/systemd/system/getty.target.wants: Directory renamed before its status could be extracted
    tar: root.x86_64/etc/systemd/user/sockets.target.wants: Directory renamed before its status could be extracted
    tar: root.x86_64/etc/systemd/user: Directory renamed before its status could be extracted
    tar: root.x86_64/etc/ssl/certs: Directory renamed before its status could be extracted
    tar: root.x86_64/var: Directory renamed before its status could be extracted
    tar: Exiting with failure status due to previous errors

This is due to the way tar keeps track of directories and the way
overlayfs presents inodes, according to https://github.com/coreos/bugs/issues/1095
Use bsdtar instead of tar in order to work around this issue.

While at it, download the signing key in the repository in order to no
longer rely on PGP keyservers, which are quite unstable.