windows/list_cmdlines: support 64-bit Windows 10
Using a remote Thread to return the result of GetCommandLineW does not
work well on 64-bit systems. Moreover on Windows 10, several UWP
applications are suspended and threads are created suspended there,
without a proper way to detect such a thing. Work around this issue by
introducing a timeout of 1 seconds.

Moreover, use NtQueryInformationProcess() to retrieve the address of the
PEB of the target process and read the command line from
PEB->ProcessParameters->CommandLine.Buffer. This is much more stable
anyway.