c/recover_segfault: no longer test on Windows
The programs using recover_segfault now crash on Windows:

    ./read_nullpage.exe
    Unknown x86_64 instruction @00007FFCEDD5430F: c4 c1 75 74 08 c5
    Running unknown instruction. Crash!
    Faulting instruction @00007FFCEDD5430F: c4 c1 75 74 08 c5 fd d7 c9 85 c9 c5 f8 77 74 1c

These are AVX-512 instructions:

   0:    c4 c1 75 74 08           vpcmpeqb ymm1,ymm1,YMMWORD PTR [r8]
   5:    c5 fd d7 c9              vpmovmskb ecx,ymm1
   9:    85 c9                    test   ecx,ecx
   b:    c5 f8 77                 vzeroupper

Microsoft has been adding support to such instructions since 2020 and
nowadays it is normal to see it used in a function as simple as strlen.

As it is quite hard to support YMM registers, skip testing on Windows.

References:
- https://devblogs.microsoft.com/cppblog/avx-512-auto-vectorization-in-msvc/
- https://revers.engineering/x86/pcmpeqb.pdf